update on test and comments
